refactor: reorganize folder structure and file names for anti-fraud module
